Fix price data types

+++ cc_idea/extractors/yahoo.py
@@ -29,13 +29,13 @@ def load_prices(symbol: str) -> DataFrame:
     # Validate and rename columns.
     columns = {
         'date': {'rename': 'Date', 'type': 'datetime64[ns]'},
-        'open': {'rename': 'Open', 'type': 'float'},
-        'high': {'rename': 'High', 'type': 'float'},
-        'low': {'rename': 'Low', 'type': 'float'},
-        'close':  {'rename': 'Close', 'type': 'float'},
-        'volume':  {'rename': 'Volume', 'type': 'int'},
-        'dividends':  {'rename': 'Dividends', 'type': 'int'},
-        'stock_splits':  {'rename': 'Stock Splits',  'type': 'int'},
+        'open': {'rename': 'Open', 'type': 'float64'},
+        'high': {'rename': 'High', 'type': 'float64'},
+        'low': {'rename': 'Low', 'type': 'float64'},
+        'close':  {'rename': 'Close', 'type': 'float64'},
+        'volume':  {'rename': 'Volume', 'type': 'int64'},
+        'dividends':  {'rename': 'Dividends', 'type': 'int64'},
+        'stock_splits':  {'rename': 'Stock Splits',  'type': 'int64'},
     }
     df = df.rename(columns={v['rename']: k for k, v in columns.items()})
     df = df.astype({k: v['type'] for k, v in columns.items()})
